Move device discovery to a separate file
